requirements.txt
astroid==2.0.4
autopep8==1.4.4
dodgy==0.1.9
entrypoints==0.3
flake8==3.7.7
imageio==2.5.0
isort==4.3.20
lazy-object-proxy==1.4.1
mccabe==0.6.1
numpy==1.16.3
pep8-naming==0.4.1
Pillow==6.0.0
prospector==1.1.6.2
pycodestyle==2.4.0
pydocstyle==3.0.0
pyflakes==1.6.0
pylama==7.7.1
pylint==2.1.1
pylint-celery==0.3
pylint-django==2.0.2
pylint-flask==0.5
pylint-plugin-utils==0.5
PyYAML==5.1
requirements-detector==0.6
setoptconf==0.2.0
six==1.12.0
snowballstemmer==1.2.1
wrapt==1.11.1
